Introducing boost 1.80.0 to gpgpu-sim source code.
Only boost/core and boost/dynamic_bitset are introduced to gpgpu-sim's
repository. As we want to allow a runtime-configurable sectored cache size,
the macro-hardcoded-size arrays used in sector_cache_block becomes
inappropriate, so does mem_access_sector_mask_t, which used to be an
alias to std::bitset<SECTOR_CHUNK_SIZE>. This commit is the first step
to replace std::bitset with boost::dynamic_bitset.

The benefit of using boost::dynamic_bitset is to keep as much code
intact as possible, since boost::dynamic_bitset shares a lot of common
APIs with std::bitset.

It's debatable whether copying (parts of) the boost library into the
simulator's repo is best practice. We could write our own
implementation of the dynamic_bitset. We could also ask the users to bring
their own boost library... But the problem is that boost does not
guarantee consistent API from version to version, and Boost's own website
recommends freezing the version of the Boost libraries used by the
project.